rename prometheus collector for better in situ documentation
