+++ internal/builtins/collector/prometheus/testdata/config_exclude_regex_invalid_setting.yaml
