+++ internal/builtins/collector/prometheus/testdata/config_exclude_regex_valid_setting.yaml
