+++ internal/builtins/collector/prometheus/testdata/config_include_regex_invalid_setting.yaml
