+++ internal/builtins/collector/prometheus/testdata/config_include_regex_valid_setting.yaml
