+++ internal/builtins/collector/prometheus/testdata/config_metrics_default_status_disabled_setting.toml
