+++ internal/builtins/collector/prometheus/testdata/config_metrics_default_status_invalid_setting.toml
