+++ internal/builtins/collector/prometheus/testdata/config_metrics_disabled_setting.yaml
