+++ internal/builtins/collector/prometheus/testdata/config_metrics_enabled_setting.yaml
