+++ internal/builtins/collector/prometheus/testdata/config_run_ttl_valid_setting.yaml
