+++ internal/builtins/collector/prometheus/testdata/no_urls.yaml
