+++ internal/builtins/collector/prometheus/testdata/valid.json
