+++ internal/builtins/collector/prometheus/vars.go
@@ -3,7 +3,7 @@
 // license that can be found in the LICENSE file.
 //
 
-package prom
+package prometheus
 
 import (
 	"fmt"
