+++ internal/builtins/configure_linux.go
@@ -9,7 +9,7 @@ package builtins
 
 import (
 	"github.com/circonus-labs/circonus-agent/internal/builtins/collector/linux/procfs"
-	"github.com/circonus-labs/circonus-agent/internal/builtins/collector/prom"
+	"github.com/circonus-labs/circonus-agent/internal/builtins/collector/prometheus"
 	appstats "github.com/maier/go-appstats"
 	"github.com/rs/zerolog/log"
 )
@@ -27,7 +27,7 @@ func (b *Builtins) configure() error {
 		b.logger.Info().Str("id", c.ID()).Msg("enabled builtin")
 		b.collectors[c.ID()] = c
 	}
-	prom, err := prom.New("")
+	prom, err := prometheus.New("")
 	if err != nil {
 		b.logger.Warn().Err(err).Msg("prom collector, disabling")
 	} else {
